Minor unrelated refactors
